backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
akshare>=1.17.0
apscheduler==3.10.4
python-dateutil==2.9.0
curl_cffi>=0.7.0
pyyaml>=6.0
requests>=2.28
yfinance>=0.2.40
# 用户认证相关依赖
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
